refactor: remove CreateVmiOnNodeLabeled
CreateVmiOnNodeLabeled function can be dropped in favor of
a new libvmi.WithNodeAffinityForLabel() function.
